Take utility functions out of string.js
